General Snake Improvments
- Snake Customise:
You are able to change the colour of your snake. You can change both the head and the body. You do this on a new page. It uses sessionStorage to work, so if you close the tab it won't work anymore.
- Added Buttons
Added buttons to get you to this page, also added code to easily change the colour.
- WASD
WASD is supported, with the arrow keys.
- Clean Up
Added more comments, deleted some bits of messy code, made it look generally nicer.